Reorganize test folders into unit and int.
