+++ server/jest.config.int.mjs
@@ -1,17 +1,14 @@
-/** @type {import('jest').Config} */
-
+// server/jest.config.int.mjs
 import { fileURLToPath } from 'url';
-
-// Get the root directory of the project by converting the URL to a file path
 const rootDir = fileURLToPath(new URL('.', import.meta.url));
 
 export default {
   rootDir,
-  testMatch: ['**/tests/**/*.int.test.js'],
   testEnvironment: 'node',
-  setupFilesAfterEnv: ['<rootDir>/tests/helpers/testEnv.js'],
-  //setupFilesAfterEnv: ['<rootDir>/tests/helpers/testEnv.js'],
-  maxWorkers: 1, // транзакции: безопаснее без параллели на одну БД
+  roots: ['<rootDir>/tests/int'],
+  testMatch: ['**/*.int.test.[jt]s?(x)'],
+  setupFilesAfterEnv: ['<rootDir>/tests/int/helpers/testEnv.js'],
+  maxWorkers: 1,               // одна БД — без параллели
   verbose: true,
   detectOpenHandles: true,
   forceExit: true,
