+++ server/tests/int/factories/user.js
@@ -1,4 +1,4 @@
-import { User } from '../../models';
+import { User } from '../../../models';
 
 export async function createUser(attrs = {}) {
   const defaults = {
