+++ server/tests/unit/validate-request.test.js
@@ -1,5 +1,5 @@
 import { jest } from '@jest/globals';
-import { validateRequest } from '../middlewares/validate-request.js';
+import { validateRequest } from '../../middlewares/validate-request.js';
 import { z } from 'zod';
 
 describe('validateRequest middleware (unit)', () => {
